requirements.txt
Flask==2.0.2
requests==2.22.0
face_recognition==1.3.0
Flask_RESTful==0.3.9
numpy==1.19.4
Pillow==8.4.0
